% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{add-file-templates}
\alias{add-file-templates}
\title{Add Templates From Paths}
\arguments{
\item{...}{specify list of templates as key-value pairs where key is the
name of the template and value is the path to the template on file.}
}
\value{
Self (invisibly)
}
\description{
Add templates to library from file paths.
}
\details{
All templates must be named.
}
\examples{
tera <- new_engine()

writeLines(
  '<p>Hello {{ x }}. This is {{ y }}.</p>',
  con = file.path(tempdir(), "hello-world.html")
)

writeLines(
  '<img src="{{ img_src }}">',
  con = file.path(tempdir(), "img-src.html")
)

tera$add_file_templates(
  "hello-world" = file.path(tempdir(), "hello-world.html"),
  "img-src" = file.path(tempdir(), "img-src.html")
)

tera
}
