% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{add_file_templates}
\alias{add_file_templates}
\title{Add File Templates}
\usage{
add_file_templates(...)
}
\arguments{
\item{...}{specify list of templates as key-value pairs where key is the
name of the template and value is the path to the template on file.}
}
\value{
Self (invisibly)
}
\description{
Use \verb{$add_file_templates()} to add templates to library from file paths.
}
\details{
All templates must be named.
}
\examples{
tera <- ExTera$new()

writeLines(
  '<p>Hello {{ x }}. This is {{ y }}.</p>',
  con = file.path(tempdir(), "hello-world.html")
)

writeLines(
  '<img src="{{ img_src }}">',
  con = file.path(tempdir(), "img-src.html")
)

tera$add_file_templates(
  "hello-world" = file.path(tempdir(), "hello-world.html"),
  "img-src" = file.path(tempdir(), "img-src.html")
)

tera
}
