% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{add_string_templates}
\alias{add_string_templates}
\title{Add String Templates}
\usage{
add_string_templates(...)
}
\arguments{
\item{...}{specify list of templates as key-value pairs where key is the
name of the template and value is a string template.}
}
\value{
Self (invisibly)
}
\description{
Use \verb{$add_string_templates()} to add templates to library from character
strings.
}
\details{
All templates must be named.
}
\examples{
tera <- ExTera$new()

tera$add_string_templates(
  "hello-world" = '<p>Hello {{ x }}. This is {{ y }}.</p>',
  "img-src" = '<img src="{{ img_src }}">'
)

tera
}
