% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{autoescape-on}
\alias{autoescape-on}
\alias{autoescape-off}
\title{Autoescaping}
\usage{
$autoescape_on()
$autoescape_off()
}
\value{
Self (invisibly)
}
\description{
Turn autoescaping of HTML on or off. Autoescaping is on by
default.
}
\details{
Autoescaping only applies to templates whose names end with ".html",
".htm", or ".xml".
}
\examples{
tera <- new_engine()

tera$add_string_templates(
  "hello-world" = '<p>Hello {{ x }}. This is {{ y }}.</p>',
  "hello-world.html" = '<p>Hello {{ x }}. This is {{ y }}.</p>'
)

# not recognized as html
tera$render_to_string(
  "hello-world",
  x = "&world",
  y = "an apostrophe, '"
)

# html
tera$render_to_string(
  "hello-world.html",
  x = "&world",
  y = "an apostrophe, '"
)

# turn off autoescape
tera$autoescape_off()

tera$render_to_string(
  "hello-world.html",
  x = "&world",
  y = "an apostrophe, '"
)
}
