% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{autoescape_off}
\alias{autoescape_off}
\title{Turn Off Autoescaping}
\usage{
autoescape_off()
}
\value{
Self (invisibly)
}
\description{
Use \verb{$autoescape_off()} to turn off autoescaping of HTML.
}
\details{
Autoescaping is on by default.
}
\examples{
tera <- ExTera$new()

tera$add_string_templates(
  "hello-world.html" = '<p>Hello {{ x }}. This is {{ y }}.</p>'
)

tera$autoescape_off()

tera$render_to_string(
  "hello-world.html",
  x = "&world",
  y = "an apostrophe, '"
)
}
