% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{autoescape_on}
\alias{autoescape_on}
\title{Turn On Autoescaping}
\usage{
autoescape_on()
}
\value{
Self (invisibly)
}
\description{
Use \verb{$autoescape_on()} to turn on autoescaping of HTML. Autoescaping is on by
default.
}
\details{
Autoescaping only applies to templates whose names end with \code{.html},
\code{.htm}, or \code{.xml}.
}
\examples{
tera <- ExTera$new()

tera$add_string_templates(
  "hello-world" = '<p>Hello {{ x }}. This is {{ y }}.</p>',
  "hello-world.html" = '<p>Hello {{ x }}. This is {{ y }}.</p>'
)

# not recognized as html
tera$render_to_string(
  "hello-world",
  x = "&world",
  y = "an apostrophe, '"
)

# html
tera$render_to_string(
  "hello-world.html",
  x = "&world",
  y = "an apostrophe, '"
)
}
