% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extera-package.R
\docType{package}
\name{extera-package}
\alias{extera}
\alias{extera-package}
\title{extera: Template Engine Inspired by 'tera'}
\description{
Template engine powered by the 'tera' 'Rust' crate and 'extendr'. Users mainly interact with an ExTera R6 object, which serves as a template library, including templates that depend on each other, a feature called in "inheritance" in tera. To render a template, users simply provide a context, or a dataset consisting of variable names and values. Template syntax supports additional logic, including built-in filters, tests, and functions, as well as loops, conditions, and inheritance. Documentation for tera can be found at \url{https://keats.github.io/tera/docs/}.
}
\author{
\strong{Maintainer}: Kenneth Blake Vernon \email{kenneth.b.vernon@gmail.com} (\href{https://orcid.org/0000-0003-0098-5092}{ORCID}) [copyright holder]

}
\keyword{internal}
