% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{list-templates}
\alias{list-templates}
\title{List Templates In Library}
\usage{
$list_templates()
}
\value{
NULL (invisibly)
}
\description{
List current templates in library.
}
\examples{
tera <- new_engine()

tera$add_string_templates(
  "hello-world" = '<p>Hello {{ x }}. This is {{ y }}.</p>',
  "img-src" = '<img src="{{ img_src }}">'
)

tera$list_templates()
}
