% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{list_templates}
\alias{list_templates}
\title{List Templates}
\usage{
list_templates()
}
\value{
character vector of template names.
}
\description{
Use \verb{$list_templates()} to list current templates in library.
}
\examples{
tera <- ExTera$new()

tera$add_string_templates(
  "hello-world" = '<p>Hello {{ x }}. This is {{ y }}.</p>',
  "img-src" = '<img src="{{ img_src }}">'
)

tera$list_templates()
}
