% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{new}
\alias{new}
\title{Initialize ExTera}
\usage{
new(dir = NULL)
}
\arguments{
\item{dir}{character scalar, a glob pattern with \code{*} wildcards indicating
a potentially nested directory containing multiple file templates. If
\code{NULL} (the default), an \code{ExTera} with an empty library is initialized.
See details for more information.}
}
\value{
Self (invisibly)
}
\description{
Use \code{Extera$new()} to create a new \code{ExTera} object. Will populate template
library with files in \code{dir} if specified.
}
\details{
The glob pattern \verb{templates/*.html} will match all files with the
.html extension located directly inside the \code{templates} folder, while the
glob pattern \verb{templates/**/*.html} will match all files with the .html
extension directly inside or in a subdirectory of \code{templates}. The
default naming convention is to give each template their full relative
path from \code{templates} or whatever the directory is called.
}
\examples{
# initialize ExTera with empty library
tera <- ExTera$new()
tera

# initialize ExTera from directory with glob
template_dir <- file.path(tempdir(), "templates")

dir.create(template_dir)

tmp <- file.path(
  template_dir,
  "hello-world-template.html"
)

writeLines(
  text = '<p>Hello {{ x }}. This is {{ y }}.</p>',
  con = tmp
)

tera <- ExTera$new(dir = file.path(template_dir, "*.html"))
tera
}
