% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{render-to-string}
\alias{render-to-string}
\title{Render to String}
\arguments{
\item{template}{character scalar, the name of the template to render.}

\item{...}{specify context as key-value pairs where key is the template
variable and value is the data to inject.}
}
\value{
a single \code{character} string.
}
\description{
R6 method for \code{ExTera}: render specified template to string.

\preformatted{$render_to_string(template, ...)}
}
\details{
All context elements must be named.
}
\examples{
tera <- new_engine()

tera$add_string_templates(
  "hello-world" = '<p>Hello {{ x }}. This is {{ y }}.</p>'
)

tera$render_to_string(
  "hello-world",
  x = "world",
  y = "ExTera"
)
}
