% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-template.R
\name{render_template}
\alias{render_template}
\title{One-Off Template Rendering}
\usage{
render_template(path, outfile = NULL, ...)
}
\arguments{
\item{path}{character scalar, path to a template file}

\item{outfile}{character scalar, the path to file where template is to
be rendered. If \code{NULL} (the default), it will render the template file to a
string in the current R session.}

\item{...}{specify context as key-value pairs where key is the template
variable and value is the data to inject.}
}
\value{
outfile (invisibly)
}
\description{
For rendering a single template file, it may be preferable to use this
one-off rendering option.
}
\details{
Requires a path to a template file, not a template string.
}
\examples{
outdir <- tempdir()

tmp <- file.path(
  outdir,
  "hello-world-template.html"
)

writeLines(
  '<p>Hello {{ x }}. This is {{ y }}.</p>',
  con = tmp
)

outfile <- file.path(
  outdir,
  "hello-world-rendered.html"
)

# render to string
render_template(
  tmp,
  x = "world",
  y = "ExTera"
)

# render to file
render_template(
  tmp,
  outfile = outfile,
  x = "world",
  y = "ExTera"
)

readLines(outfile, warn = FALSE)
}
