% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-template.R
\name{render_template_file}
\alias{render_template_file}
\title{One-Off Template Rendering}
\usage{
render_template_file(template_file, outfile, ...)
}
\arguments{
\item{template_file}{character scalar, path to a template file}

\item{outfile}{character scalar, the path to file where template is to
be rendered.}

\item{...}{specify context as key-value pairs where key is the template
variable and value is the data to inject.}
}
\value{
outfile (invisibly)
}
\description{
For rendering a single template file, it may be preferable to use this
one-off rendering option.
}
\details{
Requires a path to a template file, not a template string.
}
\examples{
outdir <- tempdir()

template_file <- file.path(
  outdir,
  "hello-world-template.html"
)

cat(
  '<p>Hello {{ x }}. This is {{ y }}.</p>',
  file = template_file
)

outfile <- file.path(
  outdir,
  "hello-world-rendered.html"
)

render_template_file(
  template_file,
  outfile,
  x = "world",
  y = "ExTera"
)

readLines(outfile, warn = FALSE)

}
