% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{render_to_string}
\alias{render_to_string}
\title{Render Template to String}
\usage{
render_to_string(template, ...)
}
\arguments{
\item{template}{character scalar, the name of the template to render.}

\item{...}{specify context as key-value pairs where key is the template
variable and value is the data to inject.}
}
\value{
Rendered string.
}
\description{
Use \verb{$render_to_string()} to render specified template to string.
}
\details{
All context elements must be named.
}
\examples{
tera <- ExTera$new()

tera$add_string_templates(
  "hello-world" = '<p>Hello {{ x }}. This is {{ y }}.</p>'
)

tera$render_to_string(
  "hello-world",
  x = "world",
  y = "ExTera"
)
}
