% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{print}
\alias{print}
\title{Print ExTera}
\arguments{
\item{n}{integer scalar, number of templates to print (default is 10L)}

\item{...}{ignored}
}
\value{
Self (invisibly)
}
\description{
Print method for \code{ExTera} object.
}
