% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{render}
\alias{render}
\title{Render to File}
\arguments{
\item{template}{character scalar, the name of the template to render.}

\item{outfile}{character scalar, the path to file where template is to
be rendered.}

\item{...}{specify context as key-value pairs where key is the template
variable and value is the data to inject.}
}
\value{
outfile (invisibly)
}
\description{
R6 method for \code{ExTera}: render specified template to file.

\preformatted{$render(template, outfile, ...)}
}
\details{
All context elements must be named.
}
\examples{
tera <- new_engine()

tera$add_string_templates(
  "hello-world" = '<p>Hello {{ x }}. This is {{ y }}.</p>'
)

outfile <- file.path(tempdir(), "rendered-hello-world.html")

tera$render(
  "hello-world",
  outfile = outfile,
  x = "world",
  y = "ExTera"
)

readLines(outfile, warn = FALSE)
}
